fix(666): Tests
